cleanup - common mapstruct classes moved to own package
