+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/AddressApiDtoMapper.java
@@ -1,7 +1,7 @@
 package club.klabis.adapters.api.mappers.members;
 
 import club.klabis.api.dto.AddressApiDto;
-import club.klabis.common.DomainToDtoMapperConfiguration;
+import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
 import club.klabis.domain.members.Address;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
