+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/MembershipSuspensionInfoToApiDtoMapper.java
@@ -2,7 +2,7 @@ package club.klabis.adapters.api.mappers.members;
 
 import club.klabis.api.dto.MembershipSuspensionInfoApiDto;
 import club.klabis.api.dto.SuspendMembershipBlockersFinanceApiDto;
-import club.klabis.common.DomainToDtoMapperConfiguration;
+import club.klabis.common.mapstruct.DomainToDtoMapperConfiguration;
 import club.klabis.domain.members.MembershipSuspensionInfo;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
