+++ backend/src/main/java/club/klabis/common/mapstruct/MapperSpringConfig.java
@@ -1,4 +1,4 @@
-package club.klabis.common;
+package club.klabis.common.mapstruct;
 
 import org.mapstruct.extensions.spring.SpringMapperConfig;
