+++ backend/src/test/java/club/klabis/tests/common/MapperTestConfiguration.java
@@ -1,6 +1,6 @@
 package club.klabis.tests.common;
 
-import club.klabis.common.MapperSpringConfig;
+import club.klabis.common.mapstruct.MapperSpringConfig;
 import org.mapstruct.extensions.spring.test.ConverterScan;
 import org.springframework.boot.test.context.TestConfiguration;
